requirements.txt
slack_sdk==3.22.0
python-dotenv==1.0.0